pyproject.toml
[project]
name = "bp_ecg_raw_extractor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=13.0",
    "zstandard>=0.22",
    "Pillow>=10.0",
    "pdfplumber>=0.11",
    "polars>=0.20",
    "pyarrow>=16.0",
    "pyiceberg[s3fs]>=0.7",
    "structlog>=24.0",
    "pydantic-settings>=2.0",
    "prometheus-client>=0.20",
    "pandera[polars]>=0.20",
]

[project.optional-dependencies]
ocr = [
    "paddlepaddle-cpu>=2.6",
    "paddleocr>=2.7",
]
ocr-gpu = [
    # Use instead of 'ocr' when a CUDA-capable GPU is available.
    # Run container with: docker run --gpus all ...
    "paddlepaddle-gpu>=2.6",
    "paddleocr>=2.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "moto[s3]>=5.0",
    "ruff>=0.4",
    "pyrefly>=0.1",
    "reportlab>=4.0",
    "numpy>=1.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bp_ecg_raw_extractor"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ANN"]
ignore = ["ANN401"]

[tool.pyrefly]
pythonVersion = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"
asyncio_mode = "auto"
